Changed: moved firebase config
