+++ apps/web/src/pages/index.tsx
@@ -1,22 +1,8 @@
 import Link from "next/link";
 import { useRouter } from "next/router";
-import "firebase/auth";
-import "firebase/storage";
-import "firebase/firestore";
-import { initializeApp } from "firebase/app";
-
-const firebaseConfig = {
-  apiKey: "AIzaSyC1u9YMqY_fG-C4xCq62hAPiRQMFK7Nt18",
-  authDomain: "blinddate-381513.firebaseapp.com",
-  projectId: "blinddate-381513",
-  storageBucket: "blinddate-381513.appspot.com",
-  messagingSenderId: "802932977941",
-  appId: "1:802932977941:web:710b86663c664be865791c",
-};
 
 export default function Main() {
   const router = useRouter();
-  const app = initializeApp(firebaseConfig);
 
   function handleClick() {
     router.push("/home");
